+++ src/overlay/RandomPeerSource.cpp
@@ -23,7 +23,7 @@ namespace stellar
 
 using namespace soci;
 PeerQuery
-RandomPeerSource::maxFailures(int maxFailures, bool requireOutobund)
+RandomPeerSource::maxFailures(size_t maxFailures, bool requireOutobund)
 {
     return {false, maxFailures,
             requireOutobund ? PeerTypeFilter::ANY_OUTBOUND
@@ -72,7 +72,7 @@ RandomPeerSource::RandomPeerSource(PeerManager& peerManager,
 
 std::vector<PeerBareAddress>
 RandomPeerSource::getRandomPeers(
-    int size, std::function<bool(PeerBareAddress const&)> pred)
+    size_t size, std::function<bool(PeerBareAddress const&)> pred)
 {
     assert(size >= 0);
     if (size == 0)
